Updates to GLM month and predictions
Chris advised that the hard coding for month proportions should be based on the matrix of 0 and 1 instead of proportion based on annual attendance. The predictions will need to utilise all the variable proportions to calculate the attendance volume for all of Scotland, before calculating the proportion specifically for a health board (eg Orkney) or other variable (eg SIMDOne).